changes to make things build on MacOS 10.9 (Xcode 6)
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/branches/Platform-8.3/src/linux@69355 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ centutils/libcentutils/ctexec.c
@@ -93,7 +93,7 @@ CTCaptureOutputWithStderrEx(
 {
     DWORD ceError = ERROR_SUCCESS;
     unsigned int buffer_size = 1024;
-    unsigned int read_size, write_size;
+    ssize_t read_size, write_size;
     int out[2];
     int pid, status;
     PSTR tempOutput = NULL;
